requirements.txt
ccxt
pandas
lxml
python-telegram-bot